model_development/requirements.txt
torch==2.3.1
torchvision==0.18.1
pillow==9.4.0
clearml==
matplotlib==
numpy==

